Refactor ClickPipe E2E into shared provision + per-source stages
