+++ crates/clickhouse-cloud-api/tests/integration/mod.rs
@@ -1,3 +1,4 @@
 #![allow(dead_code)]
 
+pub mod stages;
 pub mod support;
